webapp/requirements.txt
python-dotenv
Flask
SQLAlchemy
psycopg2-binary

